# api/openapi.yaml
openapi: 3.0.3
info:
  title: Live Streaming Tipping Platform API
  description: |
    Secure, scalable live-streaming platform where performers broadcast video, 
    viewers watch, and viewers tip virtual tokens in real time.
    
    ## Key Features
    - Real-time WebRTC streaming with SFU
    - Virtual token economy with instant tipping
    - Age verification and KYC compliance
    - Automated and human moderation
    - PCI-compliant payment processing
    
    ## Authentication
    All endpoints require JWT authentication except registration, login, and public endpoints.
    Include the JWT token in the Authorization header: `Bearer <token>`
    
    ## Rate Limiting
    - General API: 1000 requests/hour per user
    - Streaming endpoints: 100 requests/minute per user
    - Payment endpoints: 10 requests/minute per user
    - Tipping: 60 tips/minute per user
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@livepanty.com
  license:
    name: Proprietary
servers:
  - url: https://api.livepanty.com/v1
    description: Production server
  - url: https://staging-api.livepanty.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Retrieve public user profile information
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserProfile'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Users
      summary: Update user profile
      description: Update current user's profile information
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}/preferences:
    get:
      tags:
        - Users
      summary: Get user preferences
      description: Retrieve current user's preferences and settings
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User preferences retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '403':
          $ref: '#/components/responses/Forbidden'

    patch:
      tags:
        - Users
      summary: Update user preferences
      description: Update current user's preferences and settings
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'

  # ============================================================================
  # KYC & VERIFICATION ENDPOINTS
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User ID
...
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [viewer, performer, admin, moderator]
        status:
          type: string
          enum: [active, suspended, banned, pending_verification]
        country:
          type: string
        timezone:
          type: string
        date_of_birth:
          type: string
          format: date
        profile_image_url:
          type: string
          format: uri
        bio:
          type: string
        is_public:
          type: boolean
        email_verified:
          type: boolean
        two_factor_enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PublicUserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        username:
          type: string
        country:
          type: string
        profile_image_url:
          type: string
          format: uri
        bio:
          type: string
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        display_name:
          type: string
          minLength: 2
          maxLength: 100
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: "^[a-zA-Z0-9_]+$"
        bio:
          type: string
          maxLength: 500
        profile_image_url:
          type: string
          format: uri
        is_public:
          type: boolean
        timezone:
          type: string
    UserPreferences:
      type: object
      properties:
        notification_email:
          type: boolean
        notification_push:
          type: boolean
        notification_tips:
          type: boolean
        notification_followers:
          type: boolean
        language:
          type: string
          example: "en"
        currency:
          type: string
          example: "USD"
        timezone:
          type: string
          example: "UTC"
        privacy_level:
          type: string
          enum: [strict, standard, open]
        allow_private_messages:
          type: boolean
        allow_followers:
          type: boolean
    UpdatePreferencesRequest:
      type: object
      properties:
        notification_email:
          type: boolean
        notification_push:
          type: boolean
        notification_tips:
          type: boolean
        notification_followers:
          type: boolean
        language:
          type: string
        currency:
          type: string
        timezone:
          type: string
        privacy_level:
          type: string
          enum: [strict, standard, open]
        allow_private_messages:
          type: boolean
        allow_followers:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time